BTFS-1849: added file
